eve: moves redis output code to new file
