+++ src/util-logopenfile.c
@@ -35,32 +35,8 @@
 #include "util-logopenfile-tile.h"
 
 #ifdef HAVE_LIBHIREDIS
-#include <hiredis/hiredis.h>
-
-const char * redis_push_cmd = "LPUSH";
-const char * redis_publish_cmd = "PUBLISH";
-
-
-#if HAVE_LIBEVENT == 1
-
-#include <hiredis/async.h>
-#include <hiredis/adapters/libevent.h>
-
-#endif // HAVE_LIBEVENT
-
-
-typedef struct SCLogRedisContext_ {
-	redisContext *sync;
-
-#if HAVE_LIBEVENT == 1
-	redisAsyncContext *async;
-	struct event_base *ev_base;
-#endif // HAVE_LIBEVENT
-
-} SCLogRedisContext;
-
-#endif // HAVE_LIBHIREDIS
-
+#include "util-logopenfile-redis.h"
+#endif
 
 /** \brief connect to the indicated local stream socket, logging any errors
  *  \param path filesystem path to connect to
@@ -358,341 +334,6 @@ int SCConfLogReopen(LogFileCtx *log_ctx)
     return 0;
 }
 
-
-#ifdef HAVE_LIBHIREDIS
-
-/** \brief SCLogRedisContextAlloc() - Allocates and initalizes redis context
- *  \param async indicates that async mode will be used
- *  \retval SCLogRedisContext * pointer if succesful, EXIT_FAILURE program if not
- */
-static SCLogRedisContext * SCLogRedisContextAlloc(int async)
-{
-
-    SCLogRedisContext* ctx = (SCLogRedisContext*) SCMalloc(sizeof(SCLogRedisContext));
-
-	if (unlikely(ctx == NULL)) {
-        SCLogError(SC_ERR_MEM_ALLOC, "Unable to allocate redis context!");
-        exit(EXIT_FAILURE);
-	}
-
-	ctx->sync    = NULL;
-
-#if HAVE_LIBEVENT == 1
-	ctx->ev_base = NULL;
-	ctx->async   = NULL;
-
-    if (async) {
-        ctx->ev_base = event_base_new();
-        if (unlikely(ctx->ev_base == NULL)) {
-            SCLogError(SC_ERR_MEM_ALLOC, "Unable to allocate redis async event base!");
-            exit(EXIT_FAILURE);
-        }
-    }
-#endif
-
-	return ctx;
-}
-
-/** \brief SCLogRedisContextFree() free redis context
- *  \param async indicates that async mode was used
- */
-static void SCLogRedisContextFree(SCLogRedisContext *ctx, int async)
-{
-
-    if (ctx == NULL) {
-        return;
-    }
-#if HAVE_LIBEVENT == 1
-    if (async) {
-        if (ctx->ev_base != NULL) {
-            event_base_free(ctx->ev_base);
-        }
-    }
-#endif
-    SCFree(ctx);
-}
-
-/** \brief SCLogFileCloseRedis() Closes redis log more
- *  \param log_ctx Log file context allocated by caller
- */
-static void SCLogFileCloseRedis(LogFileCtx *log_ctx)
-{
-    SCLogRedisContext * ctx = log_ctx->redis;
-
-    if ( ctx == NULL) {
-        return;
-    }
-
-    /* asynchronous */
-    if (log_ctx->redis_setup.async) {
-#if HAVE_LIBEVENT == 1
-        if (ctx->async != NULL) {
-            redisAsyncFree(ctx->async);
-        }
-        if (ctx->ev_base) {
-            event_base_loopbreak(ctx->ev_base);
-        }
-        ctx->async = NULL;
-#endif
-    } else {
-
-        /* synchronous */
-        if (ctx->sync) {
-            redisReply *reply;
-            int i;
-            for (i = 0; i < log_ctx->redis_setup.batch_count; i++) {
-                redisGetReply(ctx->sync, (void **)&reply);
-                if (reply)
-                    freeReplyObject(reply);
-            }
-            redisFree(ctx->sync);
-            ctx->sync = NULL;
-        }
-        log_ctx->redis_setup.tried = 0;
-        log_ctx->redis_setup.batch_count = 0;
-    }
-}
-
-static int SCConfLogReopenRedis(LogFileCtx *log_ctx);
-
-#if HAVE_LIBEVENT == 1
-
-#if HIREDIS_MAJOR == 0 && HIREDIS_MINOR < 11
-
-/** \brief RedisConnectCallback() Closes redis log more
- *  \param c redis async context
- */
-void RedisConnectCallback(const redisAsyncContext *c)
-{
-	SCLogInfo("Connected to redis server.");
-}
-#else
-
-/** \brief RedisConnectCallback() Closes redis log more
- *  \param c redis async context
- *  \param status status reported by async caller
- */
-void RedisConnectCallback(const redisAsyncContext *c, int status)
-{
-	SCLogInfo("Connected to redis server. Status [%d]", status);
-}
-
-#endif // HIREDIS_MAJOR == 0 && HIREDIS_MINOR < 11
-
-/** \brief RedisDisconnectCallback() Callback when disconnection from redis happens.
- *  \param c redis async context
- *  \param status status reported by async caller
- */
-void RedisDisconnectCallback(const redisAsyncContext *c, int status)
-{
-	SCLogInfo("Disconnected from redis server. Status [%d]", status);
-}
-
-
-/** \brief SCRedisAsyncCommandCallback() Callback when reply from redis happens.
- *  \param c redis async context
- *  \param r redis reply
- *  \param privvata opaque datq with pointer to LogFileCtx
- */
-static void SCRedisAsyncCommandCallback (redisAsyncContext *c, void *r, void *privdata)
-{
-
-    LogFileCtx *file_ctx = privdata;
-    SCLogRedisContext *ctx = file_ctx->redis;
-
-    redisReply *reply = r;
-
-    /* Disconnection or lost reply may have happened */
-    if (reply == NULL) {
-        if (ctx->ev_base != NULL) {
-            event_base_loopbreak(ctx->ev_base);
-        }
-        SCConfLogReopenRedis(file_ctx);
-        return;
-    }
-
-    if (reply->type == REDIS_REPLY_INTEGER ) {
-        SCLogDebug("redis reply: %lld\n", reply->integer);
-    }
-
-    if (reply->type == REDIS_REPLY_STRING && reply->str) {
-        SCLogDebug("redis reply: %s\n", reply->str);
-    }
-
-    redisAsyncHandleWrite(c);
-}
-
-#endif // HAVE_LIBEVENT == 1
-
-/** \brief SCConfLogReopenRedis() Open or re-opens connection to redis for logging.
- *  \param log_ctx Log file context allocated by caller
- */
-static int SCConfLogReopenRedis(LogFileCtx *log_ctx)
-{
-	/* only try to reconnect once per second */
-	if (log_ctx->redis_setup.tried >= time(NULL)) {
-		return -1;
-	}
-
-	SCLogRedisContext * ctx = log_ctx->redis;
-	const char *redis_server = log_ctx->redis_setup.server;
-	int redis_port = log_ctx->redis_setup.port;
-
-#if HAVE_LIBEVENT == 1
-    if (log_ctx->redis_setup.async) {
-
-        if (ctx->async != NULL)  {
-            redisAsyncDisconnect(ctx->async);
-            redisAsyncFree(ctx->async);
-        }
-
-        ctx->async = NULL;
-
-        /* ASYNC */
-        redisAsyncContext *c = redisAsyncConnect(redis_server, redis_port);
-        if ( c != NULL && c->err) {
-            SCLogError(SC_ERR_SOCKET, "Error connecting to redis server: [%s] !", c->errstr);
-            redisAsyncFree(c);
-            ctx->async = NULL;
-            log_ctx->redis_setup.tried = time(NULL);
-            return -1;
-        }
-        if (c != NULL)  {
-            redisLibeventAttach(c,ctx->ev_base);
-            redisAsyncSetConnectCallback(c,RedisConnectCallback);
-            redisAsyncSetDisconnectCallback(c,RedisDisconnectCallback);
-            SCLogInfo("Connection to redis server [%s]:[%d] will use async.",
-                    redis_server, redis_port);
-
-            redisAsyncHandleWrite(c);
-
-            ctx->async = c;
-        }
-    } else
-#endif
-	{
-        /* SYNCHRONOUS */
-        if (ctx->sync != NULL)  {
-            redisFree(ctx->sync);
-        }
-
-        ctx->sync = NULL;
-
-        redisContext *c = redisConnect(redis_server, redis_port);
-        if ( c != NULL && c->err) {
-            SCLogError(SC_ERR_SOCKET, "Error connecting to redis server: [%s] !", c->errstr);
-            redisFree(c);
-            ctx->sync = NULL;
-            log_ctx->redis_setup.tried = time(NULL);
-            return -1;
-        }
-        ctx->sync = c;
-    }
-
-    log_ctx->redis = ctx;
-    log_ctx->redis_setup.tried = 0;
-    log_ctx->redis_setup.batch_count = 0;
-    return 0;
-}
-
-
-/** \brief configure and initializes redis output logginh
- *  \param conf ConfNode structure for the output section in question
- *  \param log_ctx Log file context allocated by caller
- *  \retval 0 on success
- */
-int SCConfLogOpenRedis(ConfNode *redis_node, LogFileCtx *log_ctx)
-{
-    const char *redis_server = NULL;
-    const char *redis_port = NULL;
-    const char *redis_mode = NULL;
-    const char *redis_key = NULL;
-    int async = 0;
-
-    if (redis_node) {
-        redis_server = ConfNodeLookupChildValue(redis_node, "server");
-        redis_port =  ConfNodeLookupChildValue(redis_node, "port");
-        redis_mode =  ConfNodeLookupChildValue(redis_node, "mode");
-        redis_key =  ConfNodeLookupChildValue(redis_node, "key");
-    }
-    if (!redis_server) {
-        redis_server = "127.0.0.1";
-        SCLogInfo("Using default redis server (127.0.0.1)");
-    }
-    if (!redis_port)
-        redis_port = "6379";
-    if (!redis_mode)
-        redis_mode = "list";
-    if (!redis_key)
-        redis_key = "suricata";
-    log_ctx->redis_setup.key = SCStrdup(redis_key);
-
-    ConfGetChildValueBool(redis_node, "async", &async);
-    log_ctx->redis_setup.async = async;
-
-#if HAVE_LIBEVENT == 0
-    if (async) {
-        SCLogWarning(SC_ERR_NO_LIBEVENT, "Async option not available. Compile with --enable-libevent.");
-    }
-    log_ctx->redis_setup.async = 0;
-#endif
-
-    if (!log_ctx->redis_setup.key) {
-        SCLogError(SC_ERR_MEM_ALLOC, "Unable to allocate redis key name");
-        exit(EXIT_FAILURE);
-    }
-
-    log_ctx->redis_setup.batch_size = 0;
-
-    ConfNode *pipelining = ConfNodeLookupChild(redis_node, "pipelining");
-    if (pipelining) {
-        int enabled = 0;
-        int ret;
-        intmax_t val;
-        ret = ConfGetChildValueBool(pipelining, "enabled", &enabled);
-        if (ret && enabled) {
-            ret = ConfGetChildValueInt(pipelining, "batch-size", &val);
-            if (ret) {
-                log_ctx->redis_setup.batch_size = val;
-            } else {
-                log_ctx->redis_setup.batch_size = 10;
-            }
-        }
-    }
-
-    if (!strcmp(redis_mode, "list")) {
-        log_ctx->redis_setup.command = redis_push_cmd;
-        if (!log_ctx->redis_setup.command) {
-            SCLogError(SC_ERR_MEM_ALLOC, "Unable to allocate redis key command");
-            exit(EXIT_FAILURE);
-        }
-    } else {
-        log_ctx->redis_setup.command = redis_publish_cmd;
-        if (!log_ctx->redis_setup.command) {
-            SCLogError(SC_ERR_MEM_ALLOC, "Unable to allocate redis key command!");
-            exit(EXIT_FAILURE);
-        }
-    }
-
-
-    /* store server params for reconnection */
-    log_ctx->redis_setup.server = SCStrdup(redis_server);
-    if (!log_ctx->redis_setup.server) {
-        SCLogError(SC_ERR_MEM_ALLOC, "Error allocating redis server string !");
-        exit(EXIT_FAILURE);
-    }
-    log_ctx->redis_setup.port = atoi(redis_port);
-    log_ctx->redis_setup.tried = 0;
-
-    log_ctx->redis = SCLogRedisContextAlloc(async);
-    SCConfLogReopenRedis(log_ctx);
-    log_ctx->Close = SCLogFileCloseRedis;
-
-    return 0;
-}
-
-#endif
-
 /** \brief LogFileNewCtx() Get a new LogFileCtx
  *  \retval LogFileCtx * pointer if succesful, NULL if error
  *  */
@@ -764,133 +405,6 @@ int LogFileFreeCtx(LogFileCtx *lf_ctx)
 
     SCReturnInt(1);
 }
-
-#ifdef HAVE_LIBHIREDIS
-
-/**
- * \brief LogFileWriteRedis() writes log data to redis output.
- * \param log_ctx Log file context allocated by caller
- * \param string buffer with data to write
- * \param string_len data length
- * \retval 0 on sucess;
- * \retval -1 on failure;
- */
-static int LogFileWriteRedis(LogFileCtx *file_ctx, const char *string, size_t string_len)
-{
-    if (file_ctx->redis == NULL) {
-        SCConfLogReopenRedis(file_ctx);
-        if (file_ctx->redis == NULL) {
-            return -1;
-        } else {
-            SCLogInfo("Reconnected to redis server.");
-        }
-    }
-
-	SCLogRedisContext * ctx = file_ctx->redis;
-
-#if HAVE_LIBEVENT == 1
-	/* async mode on */
-    if (file_ctx->redis_setup.async) {
-
-        redisAsyncContext * redis_async = ctx->async;
-
-        if (redis_async) {
-            SCLogDebug("redis async command: %s", file_ctx->redis_setup.command);
-            redisAsyncCommand(redis_async,
-                    SCRedisAsyncCommandCallback,
-                    file_ctx,
-                    "%s %s %s",
-                    file_ctx->redis_setup.command,
-                    file_ctx->redis_setup.key,
-                    string);
-
-            redisAsyncHandleWrite(redis_async);
-            redisAsyncHandleRead(redis_async);
-
-        }
-        return 0;
-    }
-#endif
-
-    redisContext *redis = ctx->sync;
-
-    if (unlikely(redis == NULL)) {
-        SCConfLogReopenRedis(file_ctx);
-    }
-
-    /* synchronous mode */
-    if (file_ctx->redis_setup.batch_size) {
-
-        redisAppendCommand(redis, "%s %s %s",
-                file_ctx->redis_setup.command,
-                file_ctx->redis_setup.key,
-                string);
-        if (file_ctx->redis_setup.batch_count == file_ctx->redis_setup.batch_size) {
-            redisReply *reply;
-            int i;
-            file_ctx->redis_setup.batch_count = 0;
-            for (i = 0; i <= file_ctx->redis_setup.batch_size; i++) {
-                if (redisGetReply(redis, (void **)&reply) == REDIS_OK) {
-                    freeReplyObject(reply);
-                } else {
-                    if (redis->err) {
-                        SCLogInfo("Error when fetching reply: %s (%d)",
-                                redis->errstr,
-                                redis->err);
-                    }
-                    switch (redis->err) {
-                        case REDIS_ERR_EOF:
-                        case REDIS_ERR_IO:
-                            SCLogInfo("Reopening connection to redis server");
-                            SCConfLogReopenRedis(file_ctx);
-                            if (file_ctx->redis) {
-                                SCLogInfo("Reconnected to redis server");
-                                return 0;
-                            } else {
-                                SCLogInfo("Unable to reconnect to redis server");
-                                return 0;
-                            }
-                            break;
-                        default:
-                            SCLogWarning(SC_ERR_INVALID_VALUE,
-                                    "Unsupported error code %d",
-                                    redis->err);
-                            return 0;
-                    }
-                }
-            }
-        } else {
-            file_ctx->redis_setup.batch_count++;
-        }
-    } else {
-        redisReply *reply = redisCommand(redis, "%s %s %s",
-                file_ctx->redis_setup.command,
-                file_ctx->redis_setup.key,
-                string);
-
-        /*  We may lose the reply if disconnection happens! */
-        if (reply) {
-            switch (reply->type) {
-                case REDIS_REPLY_ERROR:
-                    SCLogWarning(SC_ERR_SOCKET, "Redis error: %s", reply->str);
-                    SCConfLogReopenRedis(file_ctx);
-                    break;
-                case REDIS_REPLY_INTEGER:
-                    SCLogDebug("Redis integer %lld", reply->integer);
-                    break;
-                default:
-                    SCLogError(SC_ERR_INVALID_VALUE,
-                            "Redis default triggered with %d", reply->type);
-                    SCConfLogReopenRedis(file_ctx);
-                    break;
-            }
-        } else {
-            SCConfLogReopenRedis(file_ctx);
-        }
-    }
-    return 0;
-}
-#endif
 
 int LogFileWrite(LogFileCtx *file_ctx, MemBuffer *buffer)
 {
